re-organized filters based on feature
